fixed regex issue
The previous approach to pattern matching failed to capture the case where you guessed a word containing double letters, and one of the double letters was green or yellow while the other double letter was yellow or black.

While we're already mucking about with regex, optimized some of the patterns slightly.

Also, made string quotations consistent and moved some values to constant variables.

TODO:

`live_analysis` and `post_morten` share a bunch of duplicate code. There's an abstraction to be made here.